requirements.txt
langchain
streamlit
python-dotenv
langsmith
langchain_core
langchain_community
ollama
langserve
sse_starlette
bs4
pypdf
faiss-cpu
langchainhub
groq
cassio
beautifulsoup4
langchain-groq
sentence_transformers
PyPDF2